Break non-chainable methods out of ChainableQueries.
This is actaully quite a rewrite, becase we didn't have an instance-unique query before, so queries were compounded incorrectly. Now, a query is created for each subclass, and we reset() the query at the point of load()ing it.

The resetting stuff is a mess. Would be nice to fix.