Fix loss of configured default_locale (#130)
Some of the configuration values are being lost because
ContentfulModel::Management is created without the configured values
in some locations. ContentfulModel::Migrations::ContentTypeFactory.find,
for example, creates an instance of ContentfulModel::Management with no
configuration.

This change centralises the configuration into
ContentfulModel::Management.initialize so that all instances use the
configured values, such as default_locale.